backend/app/api/v1/endpoints/auth.py
from datetime import datetime, timedelta, timezone

from fastapi import APIRouter, HTTPException
from jose import jwt
from pydantic import BaseModel

from app.core.config import settings

router = APIRouter()

_ALGORITHM = "HS256"
_TOKEN_TTL_HOURS = 8


class LoginRequest(BaseModel):
    username: str
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    username: str


@router.post("/login", response_model=TokenResponse)
async def login(req: LoginRequest):
    if req.username != settings.admin_username or req.password != settings.admin_password:
        raise HTTPException(status_code=401, detail="Invalid credentials")

    payload = {
        "sub": req.username,
        "exp": datetime.now(timezone.utc) + timedelta(hours=_TOKEN_TTL_HOURS),
    }
    token = jwt.encode(payload, settings.secret_key, algorithm=_ALGORITHM)
    return TokenResponse(access_token=token, username=req.username)
